Set command buffers in test_controller when ALL interfaces are requested
When only the controller_interface::InterfaceConfiguration.type == ALL is given,
but the names vector is empty, the command buffer vector is created with size 0
However, when writing the commands in update(), values for all command_interfaces
are being happily read from a vector with size 0, which is not very optimal.

With this commit, we resize the command buffer on on_activate when ALL interfaces
are requested, since at that point we know all available interfaces, as they
are set by the controller_manager's switch() method.